Fix discord page build error: use server component for config loading
The discord page was using a non-existent hook (useSystemConfig).
Restructured to use:
- Server component (page.tsx) that loads config via loadSystemConfig()
- Client component (DiscordRedirect.tsx) for window.open and navigation
